@@ -16,10 +16,12 @@
   "dependencies": {
     "@commitlint/cli": "^17.6.5",
     "@commitlint/config-conventional": "^17.6.5",
-    "@emotion/react": "^11.11.1",
+    "@emotion/cache": "^11.11.0",
+    "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
     "@hookform/resolvers": "^3.1.0",
-    "@mui/material": "^5.13.4",
+    "@mui/material": "^5.15.11",
+    "@mui/material-nextjs": "^5.15.11",
     "@mui/x-date-pickers": "^6.9.2",
     "@tanstack/react-query": "^4.29.12",
     "@tanstack/react-query-devtools": "^4.29.12",
@@ -37,9 +39,9 @@
     "cookies-next": "^2.1.1",
     "date-fns": "^2.30.0",
     "eslint": "8.33.0",
-    "eslint-config-next": "13.2.0",
+    "eslint-config-next": "14.0.4",
     "eslint-plugin-simple-import-sort": "^8.0.0",
-    "next": "13.2.0",
+    "next": "14.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.44.3",
@@ -63,12 +65,13 @@
     "@storybook/nextjs": "7.5.3",
     "@storybook/react": "7.5.3",
     "@storybook/testing-library": "0.2.2",
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
-    "autoprefixer": "^10.4.14",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "autoprefixer": "^10.4.17",
     "css-loader": "^6.7.3",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-config-next": "14.1.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-config-standard-with-typescript": "^23.0.0",
     "eslint-plugin-import": "^2.27.5",
@@ -79,12 +82,12 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-storybook": "^0.6.12",
     "husky": "^8.0.3",
-    "postcss": "^8.4.24",
+    "postcss": "^8.4.35",
     "postcss-import": "^15.1.0",
     "postcss-loader": "^7.3.3",
     "storybook": "7.5.3",
     "style-loader": "^3.3.3",
-    "tailwindcss": "^3.3.2",
+    "tailwindcss": "^3.4.1",
     "tsconfig-paths-webpack-plugin": "^4.0.0",
     "typescript": "^4.9.3"
   }
