@@ -8,7 +8,7 @@ export const ENDPOINT = {
   },
   USER_MGMT: {
     USERS: '/user-management/users',
-    USERS_BY_ID: (id: string) => `/user-management/users/${id}`,
+    USERS_BY_ID: (id: string) => `user-management/users/${id}`,
     SCOPES: '/user-management/users/application-scopes',
   },
 } as const;
