@@ -1,10 +1,9 @@
-import Image from 'next/image';
+'use client';
+
 import Link from 'next/link';
 
 import { Controller } from 'react-hook-form';
 
-import logo from '@/assets/brand_light.svg';
-import loginBG from '@/assets/login-bg.png';
 import Button from '@/components/base/Button';
 import TextField from '@/components/base/Textfield';
 import Typography from '@/components/base/Typography';
@@ -26,98 +25,80 @@ const Login = () => {
   } = useLogin();
 
   return (
-    <section className="body-font font-sans overflow-hidden
-      relative h-screen bg-n-1 w-full flex items-center justify-between"
-    >
-      <div
-        className="bg-primary-600 w-[318px] h-full"
+    <div className="w-[400px]">
+      <Typography variant="headline" as="h1" size="large">
+        Sign In to
+        <br />
+        {APP_TITLE}
+      </Typography>
+      <form
+        className="w-full flex flex-col gap-8 my-8"
+        onSubmit={isSubmitting ? noop : handleSubmit(onSubmit)}
       >
-        <Image
-          src={logo}
-          alt="Brand Logo"
-          className="w-40 object-contain absolute mx-11 mt-11"
-          style={{ height: 'auto', width: '10rem' }}
+        <Controller
+          control={control}
+          name="username"
+          defaultValue=""
+          render={({ field: { onChange, value, ref } }) => (
+            <TextField
+              onChange={onChange}
+              inlineLabel="Username"
+              block
+              value={value}
+              ref={ref}
+              autoFocus
+              message={errors.username && errors.username.message}
+              error={!!errors.username}
+            />
+          )}
         />
-        <Image
-          src={loginBG}
-          alt=""
-          placeholder="blur"
-          className="h-full mx-auto"
-          style={{ objectFit: 'cover', height: 'auto', width: 'auto' }}
-        />
-      </div>
-      <div className="flex justify-center items-center grow h-full">
-        <div className="w-[400px]">
-          <Typography
-            variant="headline"
-            as="h1"
-            size="large"
-          >
-            Sign In to
-            <br />
-            {APP_TITLE}
-          </Typography>
-          <form className="w-full flex flex-col gap-8 my-8" onSubmit={isSubmitting ? noop : handleSubmit(onSubmit)}>
-            <Controller
-              control={control}
-              name="username"
-              defaultValue=""
-              render={({ field: { onChange, value, ref } }) => (
-                <TextField
-                  onChange={onChange}
-                  inlineLabel="Username"
-                  block
-                  value={value}
-                  ref={ref}
-                  autoFocus
-                  message={errors.username && errors.username.message}
-                  error={!!errors.username}
-                />
+        <Controller
+          control={control}
+          name="password"
+          defaultValue=""
+          render={({ field: { onChange, value, ref } }) => (
+            <TextField
+              autoComplete="new-password"
+              onChange={onChange}
+              inlineLabel="Password"
+              block
+              value={value}
+              ref={ref}
+              message={errors.password && errors.password.message}
+              error={!!errors.password}
+              type={showPassword ? 'text' : 'password'}
+              appendObject={(
+                <Button
+                  variant="text"
+                  className="p-0 [&>*]:fill-n-7"
+                  onClick={toggleShowPassword}
+                >
+                  {showPassword ? <IcEyeOpen /> : <IcEyesClose />}
+                </Button>
               )}
             />
-            <Controller
-              control={control}
-              name="password"
-              defaultValue=""
-              render={({ field: { onChange, value, ref } }) => (
-                <TextField
-                  autoComplete="new-password"
-                  onChange={onChange}
-                  inlineLabel="Password"
-                  block
-                  value={value}
-                  ref={ref}
-                  message={errors.password && errors.password.message}
-                  error={!!errors.password}
-                  type={showPassword ? 'text' : 'password'}
-                  appendObject={(
-                    <Button variant="text" className="p-0 [&>*]:fill-n-7" onClick={toggleShowPassword}>
-                      {showPassword ? <IcEyeOpen /> : <IcEyesClose />}
-                    </Button>
-                  )}
-                />
-              )}
-            />
-            <div className="flex justify-end">
-              <Link href="/forgot-password">
-                <Button variant="text" className="p-0 text-n-8">Forgot Password?</Button>
-              </Link>
-            </div>
-            <div className="flex justify-between items-center gap-2">
-              <Button
-                className="w-full"
-                color="primary"
-                size="small"
-                type="submit"
-                loading={isSubmitting}
-              >
-                Sign In
-              </Button>
-            </div>
-          </form>
+          )}
+        />
+        <div className="flex justify-end">
+          <Link href="/forgot-password">
+            <Button variant="text" className="p-0 text-n-8">
+              Forgot Password?
+            </Button>
+          </Link>
         </div>
-      </div>
-    </section>
+        <div className="flex justify-between items-center gap-2">
+          <Button
+            className="w-full"
+            color="primary"
+            size="small"
+            type="submit"
+            loading={isSubmitting}
+          >
+            Sign In
+          </Button>
+        </div>
+      </form>
+    </div>
   );
 };
 
